requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
selenium==4.26.1
pandas==2.2.3
python-dotenv==1.0.1
schedule==1.2.2
lxml==5.3.0
webdriver-manager==4.0.2
feedparser==6.0.11
nltk==3.9.1
textblob==0.18.0
python-dateutil==2.9.0
loguru==0.7.2
aiohttp==3.11.10
asyncio-throttle==1.0.2